Check 404 status first, then 500 if server error

+++ backend/controllers/product.controller.js
@@ -24,13 +24,17 @@ export const deleteProduct = async (req, res) => {
     const { id } = req.params; // Get product ID from request parameters
     console.log("Deleting product with id: ", id);
 
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Product not found" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         console.log("Product deleted successfully");
         res.status(200).json({ success: true, message: "Product deleted successfully" });
     }catch (error) {
         console.log("Error in deleting product: ", error.message);
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 }
